+++ packages/gatsby-transformer-cloudinary/get-image-objects.test.js
@@ -157,6 +157,16 @@ describe('getFixedImageObject', () => {
     );
   });
 
+  it('allows the user to set the width and height simultaneously', async () => {
+    const options = getDefaultOptions();
+    getPluginOptions.mockReturnValue(options);
+    const args = getDefaultArgs({ width: 100, height: 100 });
+
+    expect(await getFixedImageObject(args)).toEqual(
+      expect.objectContaining({ width: 100, height: 100 }),
+    );
+  });
+
   it('creates a srcset with multiple images based on the provided width', async () => {
     const options = getDefaultOptions();
     getPluginOptions.mockReturnValue(options);
@@ -168,28 +178,40 @@ describe('getFixedImageObject', () => {
       'https://res.cloudinary.com/cloudName/image/upload/w_320/public_id 2x',
       'https://res.cloudinary.com/cloudName/image/upload/w_480/public_id 3x',
     ];
-    expect(await getFixedImageObject(args)).toEqual(
-      expect.objectContaining({
-        srcSet: expectedSrcSet.join(','),
-      }),
+    expect((await getFixedImageObject(args)).srcSet).toEqual(
+      expectedSrcSet.join(','),
     );
   });
 
   it('creates a srcset with multiple images based on the provided height', async () => {
     const options = getDefaultOptions();
     getPluginOptions.mockReturnValue(options);
-    const args = getDefaultArgs({ height: 100 });
+    const args = getDefaultArgs({ height: 97 });
 
     const expectedSrcSet = [
-      'https://res.cloudinary.com/cloudName/image/upload/w_178/public_id 1x',
-      'https://res.cloudinary.com/cloudName/image/upload/w_267/public_id 1.5x',
-      'https://res.cloudinary.com/cloudName/image/upload/w_356/public_id 2x',
-      'https://res.cloudinary.com/cloudName/image/upload/w_533/public_id 3x',
+      'https://res.cloudinary.com/cloudName/image/upload/h_97/public_id 1x',
+      'https://res.cloudinary.com/cloudName/image/upload/h_146/public_id 1.5x',
+      'https://res.cloudinary.com/cloudName/image/upload/h_194/public_id 2x',
+      'https://res.cloudinary.com/cloudName/image/upload/h_291/public_id 3x',
     ];
-    expect(await getFixedImageObject(args)).toEqual(
-      expect.objectContaining({
-        srcSet: expectedSrcSet.join(','),
-      }),
+    expect((await getFixedImageObject(args)).srcSet).toEqual(
+      expectedSrcSet.join(','),
+    );
+  });
+
+  it('creates a srcset with multiple images based on the provided width and height', async () => {
+    const options = getDefaultOptions();
+    getPluginOptions.mockReturnValue(options);
+    const args = getDefaultArgs({ width: 89, height: 97 });
+
+    const expectedSrcSet = [
+      'https://res.cloudinary.com/cloudName/image/upload/w_89,h_97/public_id 1x',
+      'https://res.cloudinary.com/cloudName/image/upload/w_134,h_146/public_id 1.5x',
+      'https://res.cloudinary.com/cloudName/image/upload/w_178,h_194/public_id 2x',
+      'https://res.cloudinary.com/cloudName/image/upload/w_267,h_291/public_id 3x',
+    ];
+    expect((await getFixedImageObject(args)).srcSet).toEqual(
+      expectedSrcSet.join(','),
     );
   });